refactor(anta): refactor logging
